Commited a patch from Gary Ekker, with additional suggestions from Mike Kestner.
email from Mike:
                                                                                                                         
On Wed, 2005-07-06 at 14:32 -0600, Gary Ekker wrote:
> Here's a patch for gtk-sharp2-1.9.5 and gtk-sharp-1.0.8 to find the
> new version of gtkhtml in SUSE
> 10. Wade and I added it yesterday to the SUSE build system.
>
> I didn't see anything on the sourceforge project page informing me how
> to send patches. Please let me know if there is a preferred method.
                                                                                                                         
I don't have a problem with something like this going upstream to trunk,
but it should require the latest release and it should be an = check
(not >=) because if past experience is any indicator, they'll be bumping
the so version every release.
                                                                                                                         
As far as the branch goes, you should probably just patch that until
they release 3.8.0, because I'd rather not do stable releases just to
support the latest unstable gtkhtml releases.
                                                                                                                         
Wade, feel free to commit to trunk with that s/>=/= change.
                                                                                                                         


svn path=/trunk/gtk-sharp/; revision=47313