gtk: Move all .custom files to partial classes
There are no real code changes in this commit, just a lot of file
renaming and boilerplate additions.

A few .custom files are just removed, because the corresponding class in
GTK is gone, so they were not really used anymore.

Some files need to be re-indented, but that will be done in a separate
commit, so that git can track the renamed files correctly and not be
confused by all the changes.