generator: add reference to System.Core in the MSBuild csproj
Extension methods (such as the one we have in XmlElementExtensions.cs)
need to link to System.Runtime.CompilerServices.ExtensionAttribute
class in order to compile. In .NET 3.5, this was part of System.Core
(not mscorlib like in .NET 4.5) [1], so we need to link to it in order
to compile via MSBuild.

[1] http://msdn.microsoft.com/en-us/library/system.runtime.compilerservices.extensionattribute%28v=vs.90%29.aspx