cairo: Remove FormatStrideForWidth and PathExtents methods from Context
Those methods are not in Mono.Cairo, and are not necessary for Gtk#, so
I'm assuming they are not really needed.

The idea is to have the same API as Mono.Cairo shipped with mono, so if
you need those methods, get them first in Mono.Cairo available at
https://github.com/mono/mono

Also move the HasCurrentPoint property to its logical place in the
source, next to CurrentPoint.