Add method GetRawOwnedObject to Builder. This method performs a GetRawObject, but also increases the refcount.
This is (at least) needed in cases where you wish to have a class inherting a widget,
and use a builder in the constructor : base (builder.GetRawObject("name"))
Because no ref is taken the eventual freeing of the builder will cause the our object (this)
to have a refcount 0.
This is an issue since the freeing of our ToggleRef (which should happen later) performs an unref.